fix: split server/client return-to utils
