+++ packages/web/src/pages/api/auth.callback.ts
@@ -13,7 +13,7 @@ import {
   RETURN_TO_KEY,
   ROOT,
   sanitizeReturnTo,
-} from "@/lib/authentication/return-to";
+} from "../../server/auth/return-to";
 
 class NoCodeError extends TaggedError("NoCodeError") {}
 class TokenExchangeError extends TaggedError("TokenExchangeError") {}
